requirements.txt
build
coverage
deep_translator
Flask
Flask-Limiter
Flask-Testing
Flask-WTF
Gunicorn
html2text
numpy==1.21.5
pre-commit
pytest==6.2.5
pytest-mock==3.12.0
python-dotenv
requests==2.25.1
requests-mock==1.11.0
selenium==4.18.1
sentence-transformers==2.2.2
torch==2.1.2
torchaudio
torchvision
webdriver-manager==4.0.1
